Blend ggplot2 layers

Blend objects within a single layer (geom) or across multiple layers (geoms) using graphical blending modes, such as "multiply", "overlay", etc. Uses the built-in compositing support in graphical devices added in R 4.2.

blend(x, blend = "over", alpha = 1, ...)

Arguments

x

A ggplot2::Layer, such as a geom or stat, or a list of layers.

blend

The blend mode to use. The default mode, "over", corresponds to the "usual" blend mode of drawing objects on top of each other. The list of supported blend modes depends on your graphical device (see Murrell 2021), and are listed in dev.capabilities()$compositing. Blend modes can include: "clear", "source", "over", "in", "out", "atop", "dest", "dest.over", "dest.in", "dest.out", "dest.atop", "xor", "add", "saturate", "multiply", "screen", "overlay", "darken", "lighten", "color.dodge", "color.burn", "hard.light", "soft.light", "difference", and "exclusion"

Blend modes like "multiply" and "screen" are particularly useful as they are commutative: the result is the same whichever order they are applied in.

alpha

A numeric between 0 and 1 (inclusive). The opacity of a transparency mask applied to objects prior to blending.

...

Additional arguments (currently unused).

Value

An object that can be added to a ggplot() object: if the input is a ggplot2::Layer, the result is a ggplot2::Layer; if the input is a list of ggplot2::Layers, the output is a list of ggplot2::Layers.

Details

If x is a single layer / geometry and the blend_group aesthetic is not set, every graphical object (grob()) output by the geometry will be blended together using the blend blend mode. If alpha != 1, a transparency mask with the provided alpha level will be applied to each grob before blending.

If x is a single layer / geometry and the blend_group aesthetic is set, the geometry will be rendered for each subset of the data defined by the blend_group aesthetic, a transparency mask with the provided alpha level will be applied to each resulting group as a whole (if alpha != 1), then these groups will be blended together using the blend blend mode.

If x is a list of layers / geometries, those layers will be rendered separately, a transparency mask with the provided alpha level will be applied to each layer as a whole (if alpha != 1), then these layers will be blended together using the blend blend mode.

References

Murrell, Paul (2021): Groups, Compositing Operators, and Affine Transformations in R Graphics. The University of Auckland. Report. doi:10.17608/k6.auckland.17009120.v1 .