Partition a layer into subgroups (Layer operation)

A layer operation for adding a partition aesthetic to a layer.

Usage

partition(x, partition)

Arguments

x

One of:

A layer-like object: applies this operation to the layer.

A missing argument: creates an operation

Anything else: creates an operation, passing x along to the partition argument

partition

One of:

A list of quosures, such as returned by vars(), giving a (possibly multi-) column expression for the partition aesthetic. These expressions are combined using interaction() to be passed on to aes(partition = ...)

A one-sided formula, giving a single-column expression for the partition aesthetic, which is passed on to aes_(partition = ...).

Value

A layer-like object (if x is layer-like) or an operation (if not).

Details

This is a shortcut for setting the partition aesthetic of a layer.

partition(~ XXX) is roughly equivalent to adjust(aes(partition = XXX))

partition(vars(X, Y, ...)) is roughly equivalent to adjust(aes(partition = interaction(X, Y, ...)))

When a layer with a partition aesthetic is used by the following operations, the effects of the operations are applied across groups:

blend(): Blends graphical objects within the subgroups defined by the partition together using normal ("over") blending before applying its blend between subgroups.

See also

Examples

library(ggplot2)

# TODO