Layer operations

Layer operations are composable transformations that can be applied to ggplot2 layer-like objects, such as stats, geoms, and lists of stats and geoms; see the layer-like documentation page for a description of valid layer-like objects.

Usage

# S4 method for operation
show(object)

# S4 method for operation
format(x, ...)

# S4 method for adjust
format(x, ...)

# S4 method for blend
format(x, ...)

# S4 method for operation_composition
format(x, ...)

# S4 method for nop
format(x, ...)

# S4 method for operation_product
format(x, ...)

Arguments

x, object

An operation.

...

Further arguments passed to other methods.

Details

operations can be composed using the + and * operators (see operation_sum and operation_product). Addition and multiplication of operations and layer-like objects obeys the distributive law.

operations can be applied to layer-like objects using * or |>, with slightly different results:

Using *, application of operations to a list of layer-like objects is distributive. For example, list(geom_line(), geom_point()) * blend("multiply") is equivalent to list(geom_line() * blend("multiply"), geom_point() * blend("multiply")); i.e. it multiply-blends the contents of the two layers individually.

Using |>, application of operations to a list of layer-like objects is not distributive (unless the only reasonable interpretation of applying the transformation is necessarily distributive; e.g. adjust()). For example, list(geom_line(), geom_point()) |> blend("multiply") would multiply-blend both layers together, rather than multiply-blending the contents of the two layers individually.

Methods (by generic)

show: Print an operation.

format: Format an operation for printing.